# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Photo:
    properties:
      album_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
      url:
        type: string
    type: object
  models.PhotoForm:
    properties:
      album_id:
        type: integer
      description:
        maxLength: 500
        type: string
      title:
        maxLength: 100
        minLength: 2
        type: string
      url:
        type: string
    required:
    - album_id
    - title
    - url
    type: object
...
  utils.ApiRes:
    properties:
      code:
        type: integer
      data: {}
      errors: {}
      message:
        type: string
      timestamp:
        type: string
    type: object
  utils.BusinessError:
    properties:
      bizCode:
        type: integer
      details: {}
      err: {}
      httpCode:
        type: integer
    type: object
  utils.ListResponse:
    properties:
      list: {}
      total:
        type: integer
    type: object
info:
  contact: {}
  description: This is a sample go-learn server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: go-learn API
  version: "1.0"
paths:
...
  /albums/{id}/photos:
    get:
      description: 根据专辑ID获取该专辑下的所有照片，支持分页
      parameters:
      - description: 专辑ID
        in: path
        name: id
        required: true
        type: integer
      - description: 每页数量
        in: query
        name: limit
        type: integer
      - description: 偏移量
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ListResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.BusinessError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.BusinessError'
      summary: 获取专辑下的所有照片
      tags:
      - photos
...
  /photos:
    get:
      description: 获取所有照片，支持分页
      parameters:
      - description: 每页数量
        in: query
        name: limit
        type: integer
      - description: 偏移量
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ApiRes'
      summary: 获取所有照片
      tags:
      - photos
    post:
      consumes:
      - application/json
      description: 创建一条新的照片记录
      parameters:
      - description: 照片信息
        in: body
        name: photo
        required: true
        schema:
          $ref: '#/definitions/models.PhotoForm'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Photo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.BusinessError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.BusinessError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.BusinessError'
      summary: 创建照片
      tags:
      - photos
  /photos/{id}:
    delete:
      description: 根据ID删除照片
      parameters:
      - description: 照片ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.BusinessError'
      summary: 删除照片
      tags:
      - photos
    get:
      description: 根据ID获取照片详情
      parameters:
      - description: 照片ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Photo'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.BusinessError'
      summary: 获取单个照片
      tags:
      - photos
    put:
      consumes:
      - application/json
      description: 根据ID更新照片信息
      parameters:
      - description: 照片ID
        in: path
        name: id
        required: true
        type: integer
      - description: 照片信息
        in: body
        name: photo
        required: true
        schema:
          $ref: '#/definitions/models.PhotoForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.BusinessError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.BusinessError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.BusinessError'
      summary: 更新照片
      tags:
      - photos
  /photos/{id}/comments:
    get:
      description: 根据照片ID获取所有评论，支持分页
      parameters:
      - description: 照片ID
        in: path
        name: id
        required: true
        type: integer
      - description: 每页数量
        in: query
        name: limit
        type: integer
      - description: 偏移量
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ListResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.BusinessError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.BusinessError'
      summary: 获取照片的所有评论
      tags:
      - comments
...
swagger: "2.0"